update struktury
pridan model a controller pro katedry, drobne upravy kategorie, mensi provazani pomoci relations, par if v metodach, zase jsem zapomnel pravidelne commitovat

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Models\categories;
+use App\Models\department;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Foundation\Exceptions;
@@ -33,16 +34,21 @@ class CategoryController extends Controller
 
     }
 
-    function showItem(categories $name)
+    function showItem(string $short, categories $name)
     {
+        
+        $department = department::where('short', 'LIKE', $short)->first();
+        
+        if($department === null){
+            abort(404);
+        }
+        //Log::info('CategoryControler:showItem');
 
-        Log::info('CategoryControler:showItem');
-
-        $dataItems = DB::table('items')->where('categories', $name['id'])->whereNull('deleted_at')->paginate(25);
+        $dataItems = DB::table('items')->where('categories', $name['id'])->whereNull('deleted_at')->get();
         $dataLoans = DB::table('loans')->join('items', 'loans.item', '=', 'items.id')->where('categories', $name['id'])->select('loans.item', 'loans.rent_from', 'loans.rent_to')->get();
         $permition = DB::table('users')->join('permition', 'users.permition', '=', 'permition.id')->where('users.id', Auth::id())->select('permition.edit_item','permition.possibility_renting')->get();
 
-        return view('category', ['category' => $name, 'items' => $dataItems, 'loans' => $dataLoans, 'permition' => $permition]);
+        return view('category', ['category' => $name, 'items' => $dataItems, 'loans' => $dataLoans, 'permition' => $permition, 'short' => $short]);
     }
 
     
@@ -63,7 +69,7 @@ class CategoryController extends Controller
 
 
 
-    function saveCategory(Request $request)
+    function saveCategory(string $short, Request $request)
     {
         Log::info('CategoryControler:saveCategory');
 
@@ -82,15 +88,20 @@ class CategoryController extends Controller
         $category->description = is_null($request->categoryDescription) ? "": $request->categoryDescription;
         $category->save();
 
-        return redirect('categories/' . $request->categoryName);
+        return redirect('departments/'.$short.'/'.$request->categoryName);
 
     }
 
 
 
-    function addNewCategory(Request $request)
+    function addNewCategory(string $short, Request $request)
     {
         Log::info('CategoryControler:addNewCategory');
+
+        $department = department::where('short', 'LIKE', $short)->first();        
+        if($department === null){
+            abort(404);
+        }
 
         if(Auth::permition()->edit_item != 1){
             abort(403);
@@ -101,6 +112,7 @@ class CategoryController extends Controller
         $category->description = 'Zde upravte popisek kategorie';
         $check = $category->save();
         $category->name = 'Abecedně seřazená NOVÁ KATEGORIE, id: ' . $category->id;
+        $category->department_id = $department->id;
         $check1 = $category->save();
 
         return back()->withInput(array('saveCheck' => $check && $check1 ? '1' : '0'));
@@ -116,11 +128,14 @@ class CategoryController extends Controller
             return "0";
         }
 
+        $cat = categories::find($request->id);
+        if($cat === null)
+            abort(404);
+
         $data = DB::table('items')->leftJoin('loans', 'items.id', '=', 'loans.item')->leftJoin('Users', 'loans.user', '=', 'Users.id')->Join('categories', 'items.categories', '=', 'categories.id')->orderBy('categories.name', 'asc')->orderBy('items.id', 'asc')->select('Users.id as userId', 'Users.name', 'Users.surname','categories.id as categoryId', 'categories.name as categoryName',  'items.id as itemId', 'items.name as itemName' , 'loans.id', 'loans.rent_from', 'loans.rent_to')->where('categories.id', $request->id)->get();
 
         if(count($data) == 0){
             $check = DB::table('categories')->where('id', $request->id)->delete();
-
             return $check;
         }else {
